chore: cleanup and docs batched merkle tree
